0.0.4

+++ HARDWARE/SYSTICK/SysTick.h
@@ -9,6 +9,6 @@ void Delay_ms(__IO u32 nTime); // 中断的方式
 void TimingDelay_Decrement(void);
 
 void timer_init(uint32_t* pnTimer, uint32_t nValue);
-uint8_t timer_check(uint32_t nTimer);
+uint32_t timer_check(uint32_t nTimer);
 
 #endif /* __SYSTICK_H */
